+++ test/scratch_test.go
@@ -1,8 +1,11 @@
-package main
+package test
 
-import "fmt"
+import (
+	"fmt"
+	"testing"
+)
 
-func main() {
+func TestScratch(z *testing.T) {
 	var total = `          total        used        free      shared  buff/cache   available
 Mem:        7747784      130380     6942216       16624      675188     7355592
 Swap:             0           0           0
